milestone_3/requirements.txt
ansi2html==1.6.0
anyio==3.3.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
bleach==4.1.0
Brotli==1.0.9
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.0
click==8.0.3
colorama==0.4.4
comet-git-pure==0.19.16
comet-ml==3.1.4
configobj==5.0.6
cryptography==3.4.7
cycler==0.10.0
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.4.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
everett==1.0.2
Flask==2.0.2
Flask-Compress==1.10.1
idna==3.1
importlib-metadata==4.8.1
ipdb==0.13.9
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.1.0
json5==0.9.5
jsonschema
jupyter-client==7.0.5
jupyter-core==4.8.1
jupyter-dash==0.4.0
jupyter-server==1.11.0
jupyterlab==3.1.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.9
notebook==6.4.4
numpy==1.21.2
nvidia-ml-py3==7.352.0
olefile==0.46
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
pip==21.2.4
plotly==5.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.10.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.1
pyzmq==22.3.0
requests==2.26.0
requests-unixsocket==0.2.0
retrying==1.3.3
scikit-learn==1.0.1
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.8.0
setuptools==58.0.4
six==1.16.0
sniffio==1.2.0
statsmodels==0.12.2
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.7
waitress==2.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==2.0.2
wheel==0.37.0
widgetsnbextension==3.5.1
wurlitzer==1.0.3
zipp==3.5.0
